Remove duplicate files if in _build
rebar3 creates a _build subdirectory with symlinks back to src
and include files, as well as code for dependencies.

Remove duplicates in _build to avoid having to select which file to
jump to every time following a function: Emacs 25+ pops up a buffer
showing the alternatives, for each file it would otherwise show the
one in src and the same in _build.

However, don't exclude all files in _build, because we want to be
able to jump into dependencies.