Avoid symlink loops when finding files
In the helper script, add a find_files/1 function as an alternative to
filelib:wildcard/1.  The find_files/1 keeps track of visited files and
directories to avoid visiting them more than once.

Consider a symlink that point rootwards like this:

    $ ls -l some/sub/dir/
    loop -> ../..
    x.erl

The symlink structure forms a loop and the filelib:wilcard/1 will find
x.erl multiple times, and if the project is large and the symlinks
point many levels back towards the root, filelib:wildcard/1 can take
very long and consume much memory.

Btw, the filelib:wilcard/1 will indeed terminate because the kernel
will eventually return an ELOOP error when the number of followed
symlinks is too high.